feat: working cloud-based object detection stream
- Add FPS counters for encoder, detector, and stream

- Add cloud connection status and retry logic

- Fix FPS counter, shared-state, and shutdown

- Working camera -> cloud -> stream pipeline

+++ sw/cloud/main.py
@@ -48,7 +48,7 @@ async def detect(file: Annotated[UploadFile, File(...)]):
     response = {"coordinates": []}
 
     for box in results[0].boxes.xyxy:
-        response["coordinates"].append(box.tolist())
+        response["coordinates"].append(box.to(int).tolist())
 
     return response
